requirements.txt
# FolderPulse Requirements
# Core dependencies for the Python application

# GUI Framework
tkinter-dev>=8.6.0  # Usually included with Python

# Image Processing (for logos and splash screen)
Pillow>=10.0.0

# File System Monitoring
watchdog>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration Management
pyyaml>=6.0

# Logging and Utilities
colorlog>=6.7.0
click>=8.1.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Building and Packaging
pyinstaller>=5.13.0
setuptools>=68.0.0
wheel>=0.41.0

# Optional: Advanced features
# requests>=2.31.0  # For HTTP requests
# matplotlib>=3.7.0  # For data visualization
